Cast fixes
git-svn-id: https://svn.blender.org/svnroot/verse/branches/release-r6-branch@3725 49c17534-1d02-dc11-b283-0030488c597c

+++ v_util.c
@@ -34,8 +34,8 @@ void v_timer_advance(VUtilTimer *timer, double seconds)
 {
 	if(timer == NULL)
 		return;
-	timer->seconds   += (int) seconds;
-	timer->fractions += (seconds - (int) seconds) * (double) 0xffffffff;
+	timer->seconds   += (uint32) seconds;
+	timer->fractions += (uint32) ((seconds - (int) seconds) * (double) 0xffffffff);
 }
 
 double v_timer_elapsed(const VUtilTimer *timer)
